Updates and new functions
+ Electron to photon correction is now made at the “sub_backgrounds”
function (used to be in the “clean”)

+ Added the “clean_std_new”  and “_clean_std_new” functions. These will
use the average and standard deviation of the fixed energy loss cuts in
the CCD to remove cosmic rays (same as in “clean_std”). The noise input
is now an array with the noise level (mean +- noise*std) to be used in
each iteration, it seems that at least 3 iterations is necessary.

+ Added “shift_e” function to align the energy loss of every scan so
that the elastic peak is at zero. The elastic peak is fit by a “fun”
which is by default a gaussian (“gauss”), but could also be a slit
function (func = ‘slit’). For ALS data the elastic peak saturates the
detector, and the ‘slit’ function seems to be the best option.